Fixed Safari+Mobile Support + better comments
As described in the summary, I've enabled safari and mobile support for
the glow effect. Lots of feedback told me to try to add it to mobile.
Also, made some minor changes like how zoomed in the mobile page is
(made it zoomed out more) to show the traces better. Also, removed the
janky comments for better comments on the CircuitReveal.jsx file.